+++ src/windows/AzPin.Windows.Core/Models/AppRunningState.cs
@@ -3,6 +3,7 @@ namespace AzPin.Windows.Models;
 public enum AppRunningState
 {
     Unknown,
+    Fetching,
     Running,
     Stopped,
     Starting,